Create 5.2.2: Vectors Pt. 3
numEmployees is read from input as the size of the vector. Then, numEmployees elements are read from input into the vector runningLogs. Use a loop to access each element in the vector and if the element is greater than averageEmployees, output the element followed by a space.

Ex: If the input is 5 197 11 31 19 55, then the output is: